Add page not found behavior
Uses react router recommended way of adding a catch all route.
Layout now mixes in the UserStore so we are guaranteed to initialize it
even when coming in to the page not found handler directly.

Also made the logout and 'you are logged in as' bits of the main
navigation only visibile if there is a currently logged in user.