make nginx.conf configurable via helm (#182)
since we want to switch between http and https because of letsencrypt i
had to move the nginx.conf into a configmap. i took desmotes and passage
as an example because there it is already like that.

i also moved the start.sh script out of /etc/nginx/ since this is now
mounted by the configmap.