Fix release ver comparison (after 6.2.1 got introduced) (#483)
cmp() function is not optimal for the purposes we use it. Now that a patch
release got introduced, our version comparison broke as such. To fix it, change
the patch number to unlikely one in order to fix it temporarily.

Feature flagging in Happa needs a bit of refactoring in the future so that it's
easier to write, use and maintain them. For now, go with a hack.